updte

+++ app/Http/Controllers/NoteController.php
@@ -12,7 +12,10 @@ class NoteController extends Controller
      */
     public function index()
     {
-        $notes = Note::query()->orderBy('created_at', 'desc')->paginate();
+        $notes = Note::query()
+        ->where('user_id', request()->user()->id)
+        ->orderBy('created_at', 'desc')
+        ->paginate();
         return view('note.index', ['notes' => $notes]);
     }
 
@@ -33,7 +36,7 @@ class NoteController extends Controller
             'note' => ['required','string']
         ]);
 
-        $data['user_id'] = 1;
+        $data['user_id'] = $request->user()->id;
         $note = Note::create($data);
         return to_route('note.show', $note)->with('message', 'Note was created');
     }
@@ -43,6 +46,9 @@ class NoteController extends Controller
      */
     public function show(Note $note)
     {
+        if ($note->user_id !== request()->user()->id) {
+            abort(403);
+        }
         return view('note.show', ['note' => $note]);
     }
 
@@ -50,7 +56,10 @@ class NoteController extends Controller
      * Show the form for editing the specified resource.
      */
     public function edit(Note $note)
-    {
+    {   
+        if ($note->user_id !== request()->user()->id) {
+            abort(403);
+        }
         return view('note.edit', ['note' => $note]);
     }
 
@@ -59,6 +68,9 @@ class NoteController extends Controller
      */
     public function update(Request $request, Note $note)
     {
+        if ($note->user_id !== request()->user()->id) {
+            abort(403);
+        }
         $data = $request->validate([
             'note' => ['required','string']
         ]);
@@ -71,7 +83,10 @@ class NoteController extends Controller
      * Remove the specified resource from storage.
      */
     public function destroy(Note $note)
-    {
+    {   
+        if ($note->user_id !== request()->user()->id) {
+            abort(403);
+        }
         $note->delete();
 
         return to_route('note.index')->with('message', 'Note was deleted');
